+++ app/Http/Controllers/RolesController.php
@@ -8,36 +8,48 @@ use Spatie\Permission\Models\Permission;
 
 class RolesController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        $roles = Role::all();
+        if (auth()->user()->hasPermissionTo('view roles')) {
+            $roles = Role::all();
 
-        return response()->json($roles);
+            return response()->json($roles);
+        }else{
+            return response()->json([
+                'message' => 'You cannot view roles'
+            ], 403);
+        }
     }
 
     public function indexWithPermissions()
-{
-    $roles = Role::with('permissions')->get();
-
-    return response()->json($roles);
-}
+    {
+        if (auth()->user()->hasPermissionTo('view roles')) {
+            $roles = Role::with('permissions')->get();
+            return response()->json($roles);
+        }else{
+            return response()->json([
+                'message' => 'You cannot view roles'
+            ], 403);
+        }
+    }
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'name' => 'required|unique:roles,name',
-            'description' => 'nullable|string',
-        ]);
-
-        $role = Role::create($validatedData);
-
-        return response()->json($role, 201);
+        if (auth()->user()->hasPermissionTo('create roles')) {
+            $validatedData = $request->validate([
+                'name' => 'required|unique:roles,name',
+                'description' => 'nullable|string',
+            ]);
+            $role = Role::create($validatedData);
+            return response()->json($role, 201);
+        }else{
+            return response()->json([
+                'message' => 'You cannot create roles'
+            ], 403);
+        }
     }
 
     /**
@@ -45,8 +57,14 @@ class RolesController extends Controller
      */
     public function show(string $id)
     {
-        $role = Role::findOrFail($id);
-        return response()->json($role);
+        if (auth()->user()->hasPermissionTo('view roles')) {
+            $role = Role::findOrFail($id);
+            return response()->json($role);
+        }else{
+            return response()->json([
+                'message' => 'You cannot view roles'
+            ], 403);
+        }
     }
 
     /**
@@ -54,16 +72,19 @@ class RolesController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $role = Role::findOrFail($id);
-
-    $validatedData = $request->validate([
-        'name' => 'sometimes|required|unique:roles,name,' . $role->id,
-        'description' => 'nullable|string',
-    ]);
-
-    $role->update($validatedData);
-
-    return response()->json($role);
+        if (auth()->user()->hasPermissionTo('edit roles')) {
+            $role = Role::findOrFail($id);
+            $validatedData = $request->validate([
+                'name' => 'sometimes|required|unique:roles,name,' . $role->id,
+                'description' => 'nullable|string',
+            ]);
+            $role->update($validatedData);
+            return response()->json($role);
+        }else{
+            return response()->json([
+                'message' => 'You cannot edit roles'
+            ], 403);
+        }
     }
 
     /**
@@ -71,39 +92,37 @@ class RolesController extends Controller
      */
     public function destroy(string $id)
     {
-        $role = Role::findOrFail($id);
-
-    $role->delete();
-
-    return response()->json(['message' => 'Role deleted successfully']);
+        if (auth()->user()->hasPermissionTo('delete roles')) {
+            $role = Role::findOrFail($id);
+            $role->delete();
+            return response()->json(['message' => 'Role deleted successfully']);
+        }else{
+            return response()->json([
+                'message' => 'You cannot delete roles'
+            ], 403);
+        }
     }
 
     public function updatePermission(Role $role, Permission $permission)
-{
-    $grant = request()->input('grant');
-    if ($grant=="true") {
-        $role->givePermissionTo($permission);
-        return response()->json([
-            'message' => 'Permission granted successfully.'
-        ], 200);
-    } else {
-        $role->revokePermissionTo($permission);
-        return response()->json([
-            'message' => 'Permission revoked successfully.'
-        ], 200);
+    {
+        if (auth()->user()->hasPermissionTo('grant and revoke permission')) {
+            $grant = request()->input('grant');
+            if ($grant=="true") {
+                $role->givePermissionTo($permission);
+                return response()->json([
+                    'message' => 'Permission granted successfully.'
+                ], 200);
+            } else {
+                $role->revokePermissionTo($permission);
+                return response()->json([
+                    'message' => 'Permission revoked successfully.'
+                ], 200);
+            }
+        }else{
+            return response()->json([
+                'message' => 'You cannot grant and revoke permission'
+            ], 403);
+        }
     }
-}
-//     public function grantPermissions(Request $request, $roleId)
-// {
-//     $role = Role::findOrFail($roleId);
-
-//     $permissionIds = $request->input('permissions');
-
-//     $permissions = Permission::whereIn('id', $permissionIds)->get();
-
-//     $role->givePermissionTo($permissions);
-
-//     return response()->json(['message' => 'Permissions granted successfully']);
-// }
 
 }
